Reorganizing init module (first pass)
