+++ src/lib.rs
@@ -30,17 +30,8 @@
 
 #![warn(clippy::unwrap_used)]
 
-use anyhow::anyhow;
 use anyhow::Context;
-use init::init_pid1_logging;
-use init::init_rootfs;
-use init::init_syslog_logging;
-use init::network::show_network_info;
-use init::print_logo;
 use log::*;
-use netlink_packet_route::RtnlMessage;
-use rtnetlink::new_connection;
-use rtnetlink::proto::Connection;
 use sea_orm::ConnectOptions;
 use sea_orm::ConnectionTrait;
 use sea_orm::Database;
@@ -54,38 +45,18 @@ use tokio::net::UnixListener;
 use tokio_stream::wrappers::UnixListenerStream;
 use tonic::transport::{Certificate, Identity, Server, ServerTlsConfig};
 
-use crate::init::network::set_link_up;
-use crate::init::network::{add_address, add_route_v6};
-use crate::init::power::spawn_thread_power_button_listener;
-
-use ipnetwork::{IpNetwork, Ipv6Network};
-
-// use crate::init::fileio::show_dir;
 use crate::observe::observe_server::ObserveServer;
 use crate::observe::ObserveService;
 use crate::runtime::runtime_server::RuntimeServer;
 use crate::runtime::RuntimeService;
 
 mod codes;
-mod init;
+pub mod init;
 mod meta;
 mod observe;
 mod runtime;
 
 pub const AURAE_SOCK: &str = "/var/run/aurae/aurae.sock";
-pub const LOOPBACK_DEV: &str = "lo";
-
-pub const LOOPBACK_IPV6: &str = "::1";
-pub const LOOPBACK_IPV6_SUBNET: &str = "/128";
-
-pub const LOOPBACK_IPV4: &str = "127.0.0.1";
-pub const LOOPBACK_IPV4_SUBNET: &str = "/8";
-
-pub const DEFAULT_NET_DEV: &str = "eth0";
-pub const DEFAULT_NET_DEV_IPV6: &str = "fe80::2";
-pub const DEFAULT_NET_DEV_IPV6_SUBNET: &str = "/64";
-
-pub const POWER_BUTTON_DEVICE: &str = "/dev/input/event0";
 
 #[derive(Debug)]
 pub struct AuraedRuntime {
@@ -183,175 +154,6 @@ impl AuraedRuntime {
         info!("gRPC server exited successfully");
 
         Ok(())
-    }
-}
-
-#[derive(Debug)]
-pub struct SystemRuntime {
-    pub logger_level: Level,
-}
-
-impl SystemRuntime {
-    fn spawn_system_runtime_threads(&self) {
-        // ---- MAIN DAEMON THREAD POOL ----
-        // TODO: https://github.com/aurae-runtime/auraed/issues/33
-        match spawn_thread_power_button_listener(Path::new(POWER_BUTTON_DEVICE))
-        {
-            Ok(_) => {
-                info!("Spawned power button device listener");
-            }
-            Err(e) => {
-                error!(
-                    "Failed to spawn power button device listener. Error={}",
-                    e
-                );
-            }
-        }
-
-        // ---- MAIN DAEMON THREAD POOL ----
-    }
-
-    async fn configure_loopback(
-        &self,
-        handle: &rtnetlink::Handle,
-    ) -> anyhow::Result<()> {
-        if let Ok(ipv6) = format!("{}{}", LOOPBACK_IPV6, LOOPBACK_IPV6_SUBNET)
-            .parse::<IpNetwork>()
-        {
-            if let Err(e) = add_address(LOOPBACK_DEV, ipv6, handle).await {
-                return Err(anyhow!("Failed to add ipv6 address to loopback device {}. Error={}", LOOPBACK_DEV, e));
-            };
-        }
-
-        if let Ok(ipv4) = format!("{}{}", LOOPBACK_IPV4, LOOPBACK_IPV4_SUBNET)
-            .parse::<IpNetwork>()
-        {
-            if let Err(e) = add_address(LOOPBACK_DEV, ipv4, handle).await {
-                return Err(anyhow!("Failed to add ipv4 address to loopback device {}. Error={}", LOOPBACK_DEV, e));
-            }
-        };
-
-        if let Err(e) = set_link_up(handle, LOOPBACK_DEV).await {
-            return Err(anyhow!(
-                "Failed to set link up for device {}. Error={}",
-                LOOPBACK_DEV,
-                e
-            ));
-        }
-
-        Ok(())
-    }
-
-    // TODO: design network config struct
-    async fn configure_nic(
-        &self,
-        handle: &rtnetlink::Handle,
-    ) -> anyhow::Result<()> {
-        if let Ok(ipv6) =
-            format!("{}{}", DEFAULT_NET_DEV_IPV6, DEFAULT_NET_DEV_IPV6_SUBNET)
-                .parse::<Ipv6Network>()
-        {
-            if let Err(e) = add_address(DEFAULT_NET_DEV, ipv6, handle).await {
-                return Err(anyhow!(
-                    "Failed to add ipv6 address to device {}. Error={}",
-                    DEFAULT_NET_DEV,
-                    e
-                ));
-            }
-
-            if let Err(e) = set_link_up(handle, DEFAULT_NET_DEV).await {
-                return Err(anyhow!(
-                    "Failed to set link up for device {}. Error={}",
-                    DEFAULT_NET_DEV,
-                    e
-                ));
-            }
-
-            if let Ok(destv6) = "::/0".to_string().parse::<Ipv6Network>() {
-                if let Err(e) =
-                    add_route_v6(&destv6, DEFAULT_NET_DEV, &ipv6, handle).await
-                {
-                    return Err(anyhow!(
-                        "Failed to add ipv6 route to device {}. Error={}",
-                        DEFAULT_NET_DEV,
-                        e
-                    ));
-                }
-            }
-        };
-
-        Ok(())
-    }
-
-    async fn init_pid1_network(
-        &self,
-        connection: Connection<RtnlMessage>,
-        handle: &rtnetlink::Handle,
-    ) {
-        tokio::spawn(connection);
-
-        trace!("configure {}", LOOPBACK_DEV);
-        match self.configure_loopback(handle).await {
-            Ok(_) => {
-                info!("Successfully configured {}", LOOPBACK_DEV);
-            }
-            Err(e) => {
-                error!("Failed to setup loopback device. Error={}", e);
-            }
-        }
-
-        trace!("configure {}", DEFAULT_NET_DEV);
-
-        match self.configure_nic(handle).await {
-            Ok(_) => {
-                info!("Successfully configured {}", DEFAULT_NET_DEV);
-            }
-            Err(e) => {
-                error!(
-                    "Failed to configure NIC {}. Error={}",
-                    DEFAULT_NET_DEV, e
-                );
-            }
-        }
-
-        show_network_info(handle).await;
-    }
-
-    async fn init_pid1(&self) {
-        print_logo();
-
-        init_pid1_logging(self.logger_level);
-        trace!("Logging started");
-
-        trace!("Configure filesystem");
-        init_rootfs();
-
-        trace!("configure network");
-        //show_dir("/sys/class/net/", false); // Show available network interfaces
-        match new_connection() {
-            Ok((connection, handle, ..)) => {
-                self.init_pid1_network(connection, &handle).await;
-            }
-            Err(e) => {
-                error!("Could not initialize network! Error={}", e);
-            }
-        };
-
-        self.spawn_system_runtime_threads();
-
-        trace!("init of auraed as pid1 done");
-    }
-
-    fn init_pid_gt_1(&self) {
-        init_syslog_logging(self.logger_level);
-    }
-
-    pub async fn init(&self) {
-        if init::get_pid() == 1 {
-            self.init_pid1().await;
-        } else {
-            self.init_pid_gt_1();
-        }
     }
 }
